Creates componentsShared and moves relevant files.
